Bug fixes, UI fail safes, error checking

+++ client/src/components/familynews-add-edit/Item.js
@@ -83,8 +83,6 @@ const DeleteText = styled(SaveText)`
 
 const SaveEditDeleteContainer = styled.div`
     display: flex;
-    margin-left: auto;
-    margin-right: 0;
     width: fit-content;
 `;
 
@@ -188,6 +186,8 @@ const Item = ({ data }) => {
     };
 
     const onDelete = () => {
+        var r = window.confirm("Press OK to delete");
+        if (r !== true) return;
         deleteFamilyNews({ variables: { id: parseInt(data.id, 10) } });
     };
