+++ client/src/components/news-add-edit/NewsAddEditItem.js
@@ -86,8 +86,6 @@ const DeleteText = styled(SaveText)`
 
 const SaveEditDeleteContainer = styled.div`
     display: flex;
-    margin-left: auto;
-    margin-right: 0;
     width: fit-content;
 `;
 
@@ -194,6 +192,8 @@ const AddEditItem = ({ data }) => {
     };
 
     const onDelete = () => {
+        var r = window.confirm("Press OK to delete");
+        if (r !== true) return;
         deleteNews({ variables: { id: parseInt(data.id, 10) } });
     };
