+++ client/src/components/thanks-add-edit/ThanksAddEditItem.js
@@ -75,8 +75,6 @@ const DeleteText = styled(SaveText)`
 
 const SaveEditDeleteContainer = styled.div`
     display: flex;
-    margin-left: auto;
-    margin-right: 0;
     width: fit-content;
 `;
 
@@ -169,6 +167,8 @@ const AddEditItem = ({ data }) => {
     };
 
     const onDelete = () => {
+        var r = window.confirm("Press OK to delete");
+        if (r !== true) return;
         deleteNews({ variables: { id: parseInt(data.id, 10) } });
     };
